HW01

+++ main.py
@@ -2,7 +2,7 @@
 
 
 lettergrade1 = input("Enter your course 1 letter grade: ")
-credit1 = float(input("Enter your course 1 credit: "))
+credit1 = int(input("Enter your course 1 credit: "))
 
 if lettergrade1 == "A":
   gradepoint1 = 4.0
@@ -29,7 +29,7 @@ print(f"Grade point for course 1 is: {gradepoint1}")
 
 
 lettergrade2 = input("Enter your course 2 letter grade: ")
-credit2 = float(input("Enter your course 2 credit: "))
+credit2 = int(input("Enter your course 2 credit: "))
 
 if lettergrade2 == "A":
   gradepoint2 = 4.0
